requirements.txt
numpy==1.22.4
Pillow==9.4.0
PyYAML==6.0 
scikit-learn==1.2.1
scipy==1.8.1
torch==1.11.0
torchmetrics==0.11.1
torchvision==0.12.0
tqdm==4.64.1
